Rename new package.
